Move detect_cpu_topo() to lpmd_cpu.c
Move detect_cpu_topo() from lpmd_cpumask.c to lpmd_cpu.c
